Call push notification register API on every app open
