+++ ios-app/UI/MainMenuTabViewController.swift
@@ -24,6 +24,7 @@
 //
 
 import UIKit
+import Alamofire
 
 class MainMenuTabViewController: UITabBarController {
     
@@ -49,5 +50,16 @@ class MainMenuTabViewController: UITabBarController {
             viewControllers?.remove(at: 3)
         }
         
+        if ((UserDefaults.standard.string(forKey: Constants.REGISTER_FCM_TOKEN) == "true") && (UserDefaults.standard.string(forKey: Constants.REGISTER_DEVICE_TOKEN) == "true")) {
+            let deviceToken = UserDefaults.standard.string(forKey: Constants.DEVICE_TOKEN)
+            let fcmToken = UserDefaults.standard.string(forKey: Constants.FCM_TOKEN)
+            let parameters: Parameters = [
+                "device_id": deviceToken!,
+                "registration_id": fcmToken!,
+                "platform": "ios"
+            ]
+            TPApiClient.apiCall(endpointProvider: TPEndpointProvider(.registerDevice), parameters: parameters,completion: { _, _ in})
+            UserDefaults.standard.set("false", forKey: Constants.REGISTER_FCM_TOKEN)
+        }
     }
 }
